pyproject.toml
[tool.poetry]
name = "Datadynamics"
version = "0.0.1"
description = "Simulation environment for data collection."
authors = ["Lasse F. Wolff Anthony <lfwa@proton.me>"]
readme = "README.md"
license = "BSD-3-Clause"
homepage = "https://lfwa.github.io/datadynamics/"
repository = "https://github.com/lfwa/datadynamics"
exclude = ["experiments"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pettingzoo = "^1.22.2"
pygame = "^2.1.2"
numpy = "^1.24.2"
gymnasium = "^0.27.1"
matplotlib = "^3.6.3"
networkx = { extras = ["default"], version = "^3.0" }
pillow = "^9.4.0"
tqdm = "^4.65.0"
vidmaker = "^2.3.0"
h5py = "^3.8.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"
pytest = "^7.2.0"
isort = "^5.12.0"
pre-commit = "^3.1.1"
mkdocs = "^1.4.2"
mkdocs-material = "^9.0.12"
pymdown-extensions = "^9.9.2"
mkdocstrings = { extras = ["python"], version = "^0.20.0" }
importlib-metadata = "^6.0.0"
geomloss = { git = "https://github.com/lfwa/geomloss.git" }
otdd = { git = "https://github.com/lfwa/otdd.git" }


[tool.poetry.group.experiments.dependencies]
pandas = "^1.5.3"
importlib-resources = "^5.12.0"
torch = "^1.13.1"
scipy = "^1.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
